pyproject.toml
[project]
name = "hypa-thymesia"
version = "0.1.0"
description = ""
authors = [
    {name = "elilourens",email = "elilourens123@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "chromadb (>=1.0.15,<2.0.0)",
    "langgraph (>=0.5.3,<0.6.0)",
    "langchain (>=0.3.26,<0.4.0)",
    "ollama (>=0.5.1,<0.6.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "open-clip-torch (>=3.0.0,<4.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "supabase (>=2.17.0,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "python-docx (>=1.2.0,<2.0.0)",
    "pymupdf (>=1.26.3,<2.0.0)",
    "transformers (>=4.54.0,<5.0.0)",
    "torch (>=2.7.1,<3.0.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "sentence-transformers (>=5.0.0,<6.0.0)",
    "pinecone-client (>=6.0.0,<7.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
